Added structure for inputs and commands
I had to include at least one thing in the the commands control so I
just threw together the stuff for mecanum. It's nothing that is too
controversial. Uh... Yeah. Also, the convention for inputs is this:
joysticks is base level (raw stuff) (we never deal with this but it
shows up as "Joysticks to Inputs" so thought I might add it in here) ,
inputs is the stuff we get from reading joysticks ("Joysticks to Inputs"
and "Inputs to Commands"), and commands is the stuff we actually process
("Inputs to Commands"). Soooo yeah! Also sorry, fell asleep coding and
didn't sync changes. Adding stuff now... -asnanda